pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentweave-ai"
version = "0.22.0"
description = "Multi-agent AI collaboration framework (Claude, Kimi, Gemini, Codex, and more)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "gutohuida"},
]
keywords = ["claude", "kimi", "gemini", "codex", "collaboration", "agents", "ai", "code", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = []

[project.optional-dependencies]
mcp = ["fastmcp>=2.0", "pyyaml>=6.0"]
jobs = ["croniter>=2.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
docs = ["mkdocs>=1.6", "mkdocs-material>=9.5"]
all = ["fastmcp>=2.0", "croniter>=2.0", "pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/gutohuida/AgentWeave"
Documentation = "https://gutohuida.github.io/AgentWeave/"
Repository = "https://github.com/gutohuida/AgentWeave"
Issues = "https://github.com/gutohuida/AgentWeave/issues"
Changelog = "https://github.com/gutohuida/AgentWeave/blob/master/CHANGELOG.md"

[project.scripts]
agentweave = "agentweave.cli:main"
aw = "agentweave.cli:main"
agentweave-watch = "agentweave.watchdog:main"
agentweave-mcp = "agentweave.mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agentweave = ["templates/*", "templates/roles/*", "templates/skills/*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["src/agentweave/mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
